requirements.txt
aioredis==1.3.1
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
azure-core==1.30.1
azure-storage-blob==12.19.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.5
channels-redis==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==3.0.2
Django==5.0.3
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-redis==5.4.0
django-timezone-field==6.1.0
django-user-agents==0.4.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
drf-jwt==1.19.2
et-xmlfile==1.1.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
isodate==0.6.1
kombu==5.3.7
msgpack==1.0.8
numpy==1.26.4
opencv-python==4.10.0.82
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg==3.1.18
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==24.1.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.6
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
ua-parser==0.18.0
urllib3==2.2.1
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.0
zope.interface==6.4.post2
kaleido==0.2.1
plotly==5.24.1
reportlab==4.2.5
drf_yasg
django-filter
